needs to be 'long' for numRects to get a pointer to it  in our 64-bit build (why?)

+++ Source/RFBView.m
@@ -134,7 +134,7 @@
 {
     NSRect          b = [self bounds];
     const NSRect    *rects;
-    int             numRects;
+    long             numRects;
     int             i;
 
     if (drawTint)
